Eliminate floating-point math in servo control

+++ include/servo.hpp
@@ -4,13 +4,13 @@
 
 class Servo {
 public:
-    Servo() : pin(255), slice(0), top(0), attached(false), period(20000.0f) {}
+    Servo() : pin(255), slice(0), top(0), attached(false), period(20000) {}
 
-    bool attach(uint gpio_pin, float freq_hz = 50.0f, float clkdiv = 64.0f);
+    bool attach(uint gpio_pin, uint32_t freq_hz = 50, uint8_t clkdiv = 64);
     void detach();
 
     void writeMicroseconds(uint16_t us);
-    void writeAngle(float degrees);
+    void writeAngle(uint16_t degrees);
     void center();
 
 private:
@@ -18,7 +18,7 @@ private:
     uint slice;
     uint32_t top;
     bool attached;
-    float period;
+    uint32_t period; // microseconds
 
     static constexpr uint16_t min_us = 500;
     static constexpr uint16_t max_us = 2500;
